# docs/swagger.yaml
basePath: /v1
definitions:
  forms.CreateOrganizationForm:
    properties:
      description:
        maxLength: 1000
        minLength: 10
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - description
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      details: {}
      message:
        type: string
      traceId:
        type: string
      type:
        type: string
    type: object
  models.Organization:
    properties:
      createdAt:
        description: |-
          gorm:"<-:create" only allows create and read but not update
          this is avoid updating created_at with a zero value by mistake
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.PaginatedResult-models_Organization:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
      meta:
        properties:
          currentPage:
            type: integer
          nextPage:
            type: integer
          totalCount:
            type: integer
          totalPages:
            type: integer
        type: object
    type: object
...
host: localhost:9000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for the Konnect Platform
  termsOfService: http://swagger.io/terms/
  title: Konnect
  version: "1.0"
paths:
  /orgs:
    get:
      consumes:
      - application/json
      description: Get all organizations that the authenticated user belongs to
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Sort field
        enum:
        - name
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResult-models_Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's organizations
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      description: Create a new organization for the authenticated user
      parameters:
      - description: Organization data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/forms.CreateOrganizationForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - Organizations
  /orgs/{orgId}:
    delete:
      consumes:
      - application/json
      description: Delete an organization
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organizations
    get:
      consumes:
      - application/json
      description: Get a specific organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - Organizations
    put:
      consumes:
      - application/json
      description: Update an existing organization
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Organization update data
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/forms.CreateOrganizationForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organizations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"